+++ src/Runtime/ValueFetcher.php
@@ -3,6 +3,8 @@ declare(strict_types=1);
 
 namespace Remorhaz\JSON\Path\Runtime;
 
+use Generator;
+use Iterator;
 use Remorhaz\JSON\Data\Value\StructValueInterface;
 use Remorhaz\JSON\Data\Value\NodeValueInterface;
 use Remorhaz\JSON\Data\Value\ScalarValueInterface;
@@ -13,72 +15,60 @@ final class ValueFetcher implements ValueFetcherInterface
     /**
      * @param Matcher\ChildMatcherInterface $matcher
      * @param NodeValueInterface $value
-     * @return NodeValueInterface[]
+     * @return NodeValueInterface[]|Iterator
      */
-    public function fetchValueChildren(
+    public function createChildrenIterator(
         Matcher\ChildMatcherInterface $matcher,
         NodeValueInterface $value
-    ): array {
+    ): Iterator {
+        return $this->createChildrenGenerator($matcher, $value);
+    }
+
+    private function createChildrenGenerator(
+        Matcher\ChildMatcherInterface $matcher,
+        NodeValueInterface $value
+    ): Generator {
         if ($value instanceof ScalarValueInterface) {
-            return [];
+            return;
         }
 
         if ($value instanceof StructValueInterface) {
-            return $this->fetchStructChildren($value, $matcher);
+            foreach ($value->createChildIterator() as $index => $element) {
+                if ($matcher->match($index, $element, $value)) {
+                    yield $element;
+                }
+            }
+            return;
         }
 
         throw new Exception\UnexpectedNodeValueFetchedException($value);
     }
 
-    public function fetchValueDeepChildren(
+    public function createDeepChildrenIterator(
         Matcher\ChildMatcherInterface $matcher,
         NodeValueInterface $value
-    ): array {
+    ): Iterator {
+        return $this->createDeepChildrenGenerator($matcher, $value);
+    }
+
+    private function createDeepChildrenGenerator(
+        Matcher\ChildMatcherInterface $matcher,
+        NodeValueInterface $value
+    ): Generator {
         if ($value instanceof ScalarValueInterface) {
-            return [];
+            return;
         }
 
         if ($value instanceof StructValueInterface) {
-            return $this->fetchStructDeepChildren($value, $matcher);
+            foreach ($value->createChildIterator() as $index => $element) {
+                if ($matcher->match($index, $element, $value)) {
+                    yield $element;
+                }
+                yield from $this->createDeepChildrenGenerator($matcher, $element);
+            }
+            return;
         }
 
         throw new Exception\UnexpectedNodeValueFetchedException($value);
-    }
-
-    private function fetchStructChildren(NodeValueInterface $value, Matcher\ChildMatcherInterface $matcher): array
-    {
-        if (!$value instanceof StructValueInterface) {
-            // TODO: extract correct argument type?
-            throw new Exception\UnexpectedNodeValueFetchedException($value);
-        }
-        $results = [];
-        foreach ($value->createChildIterator() as $index => $element) {
-            if ($matcher->match($index, $element, $value)) {
-                $results[] = $element;
-            }
-        }
-
-        return $results;
-    }
-
-    private function fetchStructDeepChildren(NodeValueInterface $value, Matcher\ChildMatcherInterface $matcher): array
-    {
-        if (!$value instanceof StructValueInterface) {
-            // TODO: extract correct argument type?
-            throw new Exception\UnexpectedNodeValueFetchedException($value);
-        }
-
-        $results = [];
-        foreach ($value->createChildIterator() as $index => $element) {
-            if ($matcher->match($index, $element, $value)) {
-                $results[] = $element;
-            }
-            array_push(
-                $results,
-                ...$this->fetchValueDeepChildren($matcher, $element)
-            );
-        }
-
-        return $results;
     }
 }